backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncio-throttle==1.0.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
Deprecated==1.2.18
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGithub==2.6.1
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.1.1
requests==2.32.4
sniffio==1.3.1
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wrapt==1.17.2
